analyzer.py
from typing import List, Dict
from datetime import datetime, timedelta
import math
import argparse
import logging
import re
from video_extractor import VideoExtractor
import json
import os
from collections import defaultdict
from dotenv import load_dotenv

# 设置日志级别
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class VideoAnalyzer:
    def __init__(self):
        load_dotenv()
        self.api_key = os.getenv('EXA_API_KEY')
        self.extractor = VideoExtractor()
        
    def search_videos(self, keyword, platforms=None):
        """搜索视频内容"""
        if platforms is None:
            platforms = ['bilibili', 'douyin', 'kuaishou']
        
        results = {}
        for platform in platforms:
            results[platform] = self.extractor.search(platform, keyword)
        return results
    
    def analyze_data(self, data, deep_analysis=False):
        """分析视频数据"""
        analysis = {
            'total_videos': 0,
            'platform_stats': {},
            'top_videos': [],
            'trends': {}
        }
        
        # 实现数据分析逻辑
        return analysis
    
    def generate_report(self, keyword, analysis, export_format='md'):
        """生成分析报告"""
        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
        filename = f"report_{keyword}_{timestamp}.{export_format}"
        
        # 实现报告生成逻辑
        return filename

def main():
    parser = argparse.ArgumentParser(description='视频数据分析工具')
    parser.add_argument('-k', '--keyword', required=True, help='搜索关键词')
    parser.add_argument('-p', '--platforms', help='指定平台,用逗号分隔')
    parser.add_argument('--deep-analysis', action='store_true', help='启用深度分析')
    parser.add_argument('--export', default='md', help='导出格式(md/pdf)')
    
    args = parser.parse_args()
    
    analyzer = VideoAnalyzer()
    
    # 解析平台参数
    platforms = args.platforms.split(',') if args.platforms else None
    
    # 搜索视频
    results = analyzer.search_videos(args.keyword, platforms)
    
    # 分析数据
    analysis = analyzer.analyze_data(results, args.deep_analysis)
    
    # 生成报告
    report_file = analyzer.generate_report(args.keyword, analysis, args.export)
    print(f"分析报告已生成: {report_file}")

if __name__ == '__main__':
    main()
