requirements.txt
# 核心依赖
selenium>=4.15.2                # 浏览器自动化
webdriver-manager>=4.0.1       # 管理 webdriver
requests>=2.31.0               # HTTP 请求
python-dotenv>=1.0.0           # 环境变量管理

# 浏览器自动化增强
undetected-chromedriver>=3.5.3 # 反爬虫
selenium-wire>=5.1.0           # 请求拦截
fake-useragent>=1.4.0          # 随机 User-Agent

# 网页解析
beautifulsoup4>=4.12.2         # HTML 解析
lxml>=4.9.3; sys_platform == "win32"                  # Windows
lxml>=4.9.3; sys_platform == "darwin"                 # macOS
lxml>=4.9.3; sys_platform == "linux"                  # Linux

# 视频处理
yt-dlp>=2023.12.30            # 视频信息提取

# API 客户端
exa-py>=1.0.6                 # Exa.ai API
httpx>=0.25.2                 # HTTP 客户端

# 工具类
python-dateutil>=2.8.2        # 日期处理
tenacity>=8.2.3               # 重试机制
loguru>=0.7.2                 # 日志管理

# 数据处理
pandas>=2.1.3; sys_platform == "win32"                # Windows
pandas>=2.1.3; sys_platform == "darwin"               # macOS
pandas>=2.1.3; sys_platform == "linux"                # Linux
numpy>=1.26.2                 # 数值计算

# 浏览器自动化（可选）
playwright>=1.49.0; sys_platform == "win32"           # Windows
playwright>=1.49.0; sys_platform == "darwin"          # macOS
playwright>=1.49.0; sys_platform == "linux"           # Linux 